+++ modules/School_Setup/Menu.php
@@ -5,7 +5,7 @@
  * @uses $menu global var
  *
  * @see  Menu.php in root folder
- * 
+ *
  * @package RosarioSIS
  * @subpackage modules
  */
@@ -14,9 +14,7 @@ $menu['School_Setup']['admin'] = array(
 	'title' => _( 'School Setup' ),
 	'default' => 'School_Setup/Calendar.php',
 	'School_Setup/PortalNotes.php' => _( 'Portal Notes' ),
-	//FJ Portal Polls
 	'School_Setup/PortalPolls.php' => _( 'Portal Polls' ),
-	//FJ add Database Backup
 	'School_Setup/MarkingPeriods.php' => _( 'Marking Periods' ),
 	'School_Setup/Calendar.php' => _( 'Calendars' ),
 	'School_Setup/Periods.php' => _( 'Periods' ),
@@ -27,9 +25,7 @@ $menu['School_Setup']['admin'] = array(
 	'School_Setup/Schools.php' => _( 'School Information' ),
 	'School_Setup/Schools.php&new_school=true' => _( 'Add a School' ),
 	'School_Setup/CopySchool.php' => _( 'Copy School' ),
-	//FJ add School Fields
 	'School_Setup/SchoolFields.php' => _( 'School Fields' ),
-	//FJ add School Configuration
 	'School_Setup/Configuration.php'  => _( 'School Configuration' ),
 );
 
@@ -39,19 +35,21 @@ $menu['School_Setup']['teacher'] = array(
 	'School_Setup/Schools.php' => _( 'School Information' ),
 	'School_Setup/MarkingPeriods.php' => _( 'Marking Periods' ),
 	'School_Setup/Calendar.php' => _( 'Calendars' ),
-	//FJ add Periods to teachers
-	'School_Setup/Periods.php' => _( 'Periods' )
+	// Add Periods to teachers.
+	'School_Setup/Periods.php' => _( 'Periods' ),
 );
 
 $menu['School_Setup']['parent'] = array(
 	'title' => _( 'School' ),
 	'default' => 'School_Setup/Calendar.php',
 	'School_Setup/Schools.php' => _( 'School Information' ),
-	'School_Setup/Calendar.php' => _( 'Calendars' )
+	// Add Marking Periods to parents & students.
+	'School_Setup/MarkingPeriods.php' => _( 'Marking Periods' ),
+	'School_Setup/Calendar.php' => _( 'Calendars' ),
 );
 
 $exceptions['School_Setup'] = array(
 	'School_Setup/PortalNotes.php' => true,
 	'School_Setup/Schools.php&new_school=true' => true,
-	'School_Setup/Rollover.php' => true
+	'School_Setup/Rollover.php' => true,
 );