+++ CHANGES.md
@@ -10,6 +10,7 @@ Changes in 3.0
 - Add link to RosarioSIS Forum to Resources in rosariosis.sql
 - Current tab in bold in Configuration.php & Food Service module wide
 - Fix do not show Delete prompt when reloading page in Schools.php
+- Add Marking Periods to parents & students in School_Setup/Menu.php & rosariosis.sql
 
 Changes in 2.9.15
 -----------------